Added full trace finding and output
Added function to find the full trace using the flat field and initial guesses for the center at the start of the order trace ("right" side of echellogram near overscan).

Also added a function to recenter the trace, which is used both in finding the order starts and generating the full trace.

Output the fitted trace as a FITS file with some header keywords and history added.