Update extract_spectrum.py
Started the extraction module script.

Through the step of pulling out the order block image from the full frame image. This has been given its own function, for ease in the future to change how this is down (e.g. the whole interpolate or not interpolate and resulting sawtooth trace middle deal).