Added optimal extraction function
Added the function to perform optimal extraction.

It's a bit of a behemoth and maybe should be split into some slightly smaller functions. But I think it's okay for now.

Has the ability to either fit for the background of each pixel slice (like the old pipeline version) or have it fixed based on fits to the counts in the top and bottom edges of the order block image. The latter should be default (and is what I mostly tested with).