Add wavelength solution fitting
Added the function to perform the iterative sigma-clipping polynomial fitting for the wavelength solution.

This could be folded into the utils "polynomial_fit_sigma_reject", but for now I kept this separate since I wanted explicit control over keeping a record the sequence of data being fit in each iteration for plotting purposes.