Added wavecal order offset function
Added a function to calculate if there is an order offset to be applied to the preliminary wavelength solution.

In my Tull reduction for the H-alpha setup, the preliminary wavelength solution has more orders than observed so the order index doesn't necessarily match between the observed spectrum and reference. This function calculates the order offset that provides the minimum residual with a reference ThAr spectrum.