Add wavesol post process and calibrate
Added a function to "post process" the wavelength solution by looking at the polynomial coefficients as a function of order. It doesn't quite work well so I'm leaving it but not using it.

Also added a "main script" function to wavelength calibrate the object spectra, by interpolating the reference wavlength solutions.