Plotting memory leak fix
Fixed a bug that was causing a memory leak (and thus problem's finishing the pipeline) in the wavelength calibration plotting functions.

The figures weren't actually being cleared (seems like this is a common matplotlib problem). Perhaps the plots should just not be made? But for now the memory leak has been fixed.