Bug fix and added compiled csv
Fixed a bug that was resulting in all nan RV samples from failed Gaussian fitting of the combined BF. The maximum number of function evaluations for curve_fit was not high enough.

Also added a function to read in science frames and create a compiled RV csv file for the night. This is for easy access to the RV values