Fixed reduction utils import
Fixed the importing of reduction_utils, it needed to be called from . import reduction_utils. Some sort of path reference or something I guess.

Also added a catch for the wavelength solution fitting iteration that was causing a problem if lines were rejected but resulted in too few lines. Will have to actually fix this in the future