Changed docs to use autoapi
Was struggling a bit getting the API documentation right so I changed to using autoapi. Not the best but more like what I want.

The order of the modules under API Reference in the table of contents was seemingly random, so I manually turned off the autoapi add to index and created my own api_index page.