Added stellar spectrum model for BF
Added to the data directory a CSV file of a PHOENIX stellar spectrum model for using as a template with the broadening functions.

It has been modified from the raw PHOENIX model by: including only wavelengths between 3000 and 11000 angstrom, and re-sampling it to a coarser wavelength grid to decrease size.