Added init files and other path changes
Added __init__ files to the main folder and modules subdirectory. I think it's perhaps not quite best practice to do it like this, but I want an __init__ in the main folder so that I can flexibly point to the data directory hosted in tull_coude_reduction.

Also changed reference to that data directory (reference files) to be created in the pipeline run and not set in the config.