Add catch for bad order wavelength fitting
Added a catch in the case that the initial peak finding for arc lamp lines in an order finds fewer orders than the polynomial order of the wavelength solution.

Before it would break, but now there is an if/else checking for this and making sure it doesn't break.

There probably are other issues this might cause though... so this is in progress.