Fix continuum spline knots
Fixed an issue where the spline continuum fitting was breaking because of the input knot array. It seems like a rare issue where if the right edge points of the spectrum are rejected relative to the knot spacing, it breaks (something about range of knots I presume). This redefines the knot spacing for each iteration of the spline fitting.