Update extraction for binning
Updated extraction to allow 2x2 binned data. Not much had to change, but the initial guesses for the fitting of pixel slice cross dispersion profiles were updated (and are now generalized to work either with 2x2 binning or not).

NOTE: background subtraction/estimation might be a bit off in the 2x2 binning case. No change was made for that.