Add try/except to optimal extraction
Added a try/except for the optimal extraction, to break and proceed with sum extraction if optimal extraction fails on even just one order of a frame.

Surprised this hadn't come up earlier, came across this with a frame after many many nights reduced no problem. Not quite sure what the issue was, but the spatial profile fitting was seeming to fail.

This fixes the (seemingly) very rare issue.